% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratioTestStat.R
\name{asymptoticBias}
\alias{asymptoticBias}
\title{Asymptotic Bias of Ratio Test Staistic}
\usage{
asymptoticBias(gamma1, gamma2)
}
\arguments{
\item{gamma1}{p/n1 where n1 is the number of time points before potential changepoint}

\item{gamma2}{p/n2 where n2 is the number of time points after potential changepoint}
}
\value{
A numeric of the asymptotic bias
}
\description{
Calculates the asymptotic bias of the Ratio test statistic used to generate the normalized test statistic
}
\seealso{
\code{\link{ratioTestStat}}
}
\keyword{internal}
