% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratioTestStat.R
\name{asymptoticVariance}
\alias{asymptoticVariance}
\title{Asymptotic Variance of Ratio Test Statistic}
\usage{
asymptoticVariance(gamma1, gamma2)
}
\arguments{
\item{gamma1}{p/n1 where n1 is the number of time points before potential changepoint}

\item{gamma2}{p/n2 where n2 is the number of time points after potential changepoint}
}
\value{
A numeric of the asymptotic variance
}
\description{
Calculates the asymptotic variance of the Ratio test statistic used to generate the normalized test statistic
}
