% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratioTestStat.R
\name{calculateExpectedTrace}
\alias{calculateExpectedTrace}
\title{Expected Trace Calulator}
\usage{
calculateExpectedTrace(gamma1, gamma2)
}
\arguments{
\item{gamma1}{p/n1 where n1 is the number of time points before potential changepoint}

\item{gamma2}{p/n2 where n2 is the number of time points after potential changepoint}
}
\value{
A list where the first element is the integral and second element is the error.
}
\description{
Calculates the expected trace used in the test statistic
}
