% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericErrorChecks.R
\name{classErrorChecks}
\alias{classErrorChecks}
\title{Class error checking}
\usage{
classErrorChecks(Class)
}
\arguments{
\item{Class}{Logical. If TRUE then an S4 class is returned. If FALSE the estimated changepoints are returned.}
}
\description{
DEVELOPER USE ONLY. This function checks the user input for the class argument for all methods.
}
