% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cptCovClass.R
\docType{class}
\name{cptCov-class}
\alias{cptCov-class}
\alias{summary,cptCov-method}
\alias{show,cptCov-method}
\title{An S4 class for a covariance changepoint object}
\usage{
\S4method{summary}{cptCov}(object)

\S4method{show}{cptCov}(object)
}
\description{
An S4 class for a covariance changepoint object
}
\section{Methods (by generic)}{
\itemize{
\item \code{summary}: Summarises the cptCov object

Summaries the results of the covariance changepoint analysis contained in the cptCov object.

\item \code{show}: Shows the cptCov object

Shows the cptCov object and all the slots it contains. Also displays the summary of the object.
}}

\section{Slots}{

\describe{
\item{\code{data}}{An n by p matrix of the data}

\item{\code{cpts}}{A numeric vector containing the identified changepoints}

\item{\code{method}}{Character containing covariance changepoint method used}

\item{\code{noCpts}}{Either 'AMOC' for at most one changepoint; 'BinSeg' for a binary segmentation approach to dect multiple changepoints; or a positive integer specifying the number of changes}

\item{\code{testStat}}{Numeric containing the test statistic for each changepoint}

\item{\code{threshold}}{Character containing the method used for generating the threshold}

\item{\code{thresholdValue}}{Numeric value of threshold}

\item{\code{msl}}{Numeric containing the minimum segment length between changepoints}

\item{\code{date}}{Creation date of the object}

\item{\code{version}}{Version of the cpt.covariance used}
}}

