% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cptCovarianceClass.R
\docType{class}
\name{cptCovariance-class}
\alias{cptCovariance-class}
\alias{summary,cptCovariance-method}
\alias{show,cptCovariance-method}
\alias{plot,cptCovariance-method}
\alias{covEst,cptCovariance-method}
\alias{covEst}
\alias{subspaceEst,cptCovariance-method}
\alias{subspaceEst}
\title{cptCovariance: an \code{S4} class for a covariance changepoint object}
\usage{
\S4method{summary}{cptCovariance}(object)

\S4method{show}{cptCovariance}(object)

\S4method{plot}{cptCovariance}(object, x = c(), y = c())

\S4method{covEst}{cptCovariance}(object)

\S4method{subspaceEst}{cptCovariance}(object)
}
\arguments{
\item{object}{An object of S4 class \code{\linkS4class{cptCovariance}}}

\item{x}{x}

\item{y}{y}
}
\description{
Contains data and information required for further changepoint analysis,
summaries and plotting.
}
\section{Methods (by generic)}{
\itemize{
\item \code{summary}: Summarises the cptCovariance object

\item \code{show}: Shows the cptCovariance object

\item \code{plot}: Plotting method for cptCovariance object. Returns a \code{\link[ggplot2]{ggplot}} object which can be manipulated as required

\item \code{covEst}: Returns covariance estimates for each segment

\item \code{subspaceEst}: Returns a basis of the subspace estimates for each segment
}}

\section{Slots}{

\describe{
\item{\code{dataset}}{An n by p matrix of the data.}

\item{\code{cpts}}{A numeric vector containing the identified changepoints.}

\item{\code{method}}{Character containing covariance changepoint method used.}

\item{\code{numCpts}}{Either 'AMOC' for at most one changepoint; 'BinSeg' for a binary segmentation approach to detect multiple changepoints; or a positive integer specifying the number of changepoints.}

\item{\code{cptsSig}}{Data frame containing the changepoint locations along with their associated test statistic; threshold; and whether or not they were deemed significant.}

\item{\code{threshold}}{Character containing the method used for generating the threshold.}

\item{\code{thresholdValue}}{Threshold value used to determine significant changepoints. If permutation test within subspace method is used then a vector of threshold values is returned.}

\item{\code{msl}}{Minimum segment length between changepoints.}

\item{\code{subspaceDim}}{Assumed subspace dimension. Only used for Subspace method.}

\item{\code{nperm}}{Numeric value of number of permutations used in permutation test. Only used for subspace method when threshold is "PermTest".}

\item{\code{LRCov}}{Character describing the long-run covariance estimator used or a matrix containing the long-run covariance estimate. Only used for CUSUM method.}

\item{\code{covEst}}{List containing the sample covariance for each segment.}

\item{\code{subspaceEst}}{List containing a basis of the subspace for each segment. Only used for Subspace method.}

\item{\code{date}}{Creation date of the object.}

\item{\code{version}}{Version of the cpt.covariance package used.}
}}

\examples{
##
out <- new('cptCovariance',
     dataset=matrix(rnorm(300),ncol=3),
		cpts=c(50,100),
		method='Ratio',
		numCpts='AMOC',
		cptsSig=data.frame('cpts'=50,'T'=33.3,'thresholdValue'=30,significant=TRUE),
		threshold='Manual',
		thresholdValue=30,
		msl=20)
summary(out)
show(out)
plot(out)
cptsSig(out)
covEst(out)

}
