% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cptCovarianceClass.R
\name{cptCovariance-methods}
\alias{cptCovariance-methods}
\alias{summary,cptCovariance-method}
\alias{show,cptCovariance-method}
\alias{plot,cptCovariance-method}
\alias{covEst,cptCovariance-method}
\alias{covEst}
\alias{subspaceEst,cptCovariance-method}
\alias{subspaceEst}
\title{cptCovariance Methods}
\usage{
\S4method{summary}{cptCovariance}(object)

\S4method{show}{cptCovariance}(object)

\S4method{plot}{cptCovariance}(object, x = c(), y = c())

\S4method{covEst}{cptCovariance}(object)

\S4method{subspaceEst}{cptCovariance}(object)
}
\arguments{
\item{object}{An object of S4 class \code{\linkS4class{cptCovariance}}}

\item{x}{x}

\item{y}{y}
}
\description{
Methods for objects with S4 class \code{\linkS4class{cptCovariance}}
}
\section{Functions}{
\itemize{
\item \code{summary,cptCovariance-method}: Summarises the cptCovariance object

\item \code{show,cptCovariance-method}: Shows the cptCovariance object

\item \code{plot,cptCovariance-method}: Plotting method for cptCovariance object. Returns a \code{\link[ggplot2]{ggplot}} object which can be manipulated as required

\item \code{covEst,cptCovariance-method}: Returns covariance estimates for each segment

\item \code{subspaceEst,cptCovariance-method}: Returns a basis of the subspace estimates for each segment
}}

\examples{
# Create new cptCovariance object
out <- new('cptCovariance',
     dataset=matrix(rnorm(300),ncol=3),
		cpts=c(50,100),
		method='Ratio',
		numCpts='AMOC',
		cptsSig=data.frame('cpts'=50,'T'=33.3,'thresholdValue'=30,significant=TRUE),
		threshold='Manual',
		thresholdValue=30,
		msl=20)

# Summarize cptCovariance object
summary(out)

# Show cptCovariance object
show(out)

# Plot cptCovariance object
plot(out)

# Show significant changepoints
cptsSig(out)

# Estimate covariance matrices in each segment
covEst(out)

# Create new cptCovariance object with subspace method
out2 <- new('cptCovariance',
     dataset=matrix(rnorm(300),ncol=3),
		cpts=c(50,100),
		method='Subspace',
		numCpts='AMOC',
		cptsSig=data.frame('cpts'=50,'T'=33.3,'thresholdValue'=30,significant=TRUE),
		threshold='Manual',
		thresholdValue=30,
		msl=20,
		subspaceDim=1)

# Estimate subspace in each segment
subspaceEst(out2)
}
