% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cptCovSubspaceClass.R
\docType{class}
\name{cptCovSubspace-class}
\alias{cptCovSubspace-class}
\alias{show,cptCovSubspace-method}
\title{Extension of S4 cptCov class for the Subspace method}
\usage{
\S4method{show}{cptCovSubspace}(object)
}
\description{
This S4 class extends the \code{\link{cptCov}} for detailed analysis of results from the Subspace method. NOTE only methods and slots that aren't included in the \code{\link{cptCov}} are described here. Your first port of call should be the \code{\link{cptCov}} documentation.
}
\section{Methods (by generic)}{
\itemize{
\item \code{show}: Shows the cptCovSubspace object

Shows the cptCov object and all the slots it contains. Also displays the summary of the object.
}}

\section{Slots}{

\describe{
\item{\code{q}}{Assumed dimension of latent subspace}

\item{\code{nperm}}{Number of permutations used in the permutation test if this was the chosen method for calculating the threshold.}
}}

