% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cptCUSUM.R
\name{cptCUSUM}
\alias{cptCUSUM}
\title{CUSUM based covariance changepoint detection}
\usage{
cptCUSUM(
  X,
  threshold = "Asymptotic",
  numCpts = "AMOC",
  msl = 2 * ncol(X),
  LRCov = "Bartlett",
  thresholdValue = 0.05,
  errorCheck = TRUE,
  Class = TRUE
)
}
\arguments{
\item{X}{Data matrix of dimension n by p.}

\item{threshold}{Threshold choice for determining significance of changepoints. Choices include:
\itemize{
\item "Asymptotic" - Uses the asymptotic threshold derived for each method. For method="Ratio" the threshold is log(n). For method="CUSUM" the threshold is the specified quantile of the standard Normal distribution. The quantile is set by the argument thresholdValue.
\item "Manual"- A user chosen threshold which is contained in the thresholdValue argument. NOTE the normalized test statistics will be compared to the set thresholds - see details for more information.
}
If numCpts is numeric then the threshold is not used as the number of changepoints is known.}

\item{numCpts}{Number of changepoints in the data. Choices include:
\itemize{
\item "AMOC" - At Most One Changepoint; test to see if the data contains a single changepoint or not.
\item "BinSeg"- Binary segmentation is performed to detect multiple changepoints.
\item Numeric - User specified number of changepoints.
}}

\item{msl}{Minimum segment length allowed between the changepoints. NOTE this should be greater than or equal to p, the dimension of the time series.}

\item{LRCov}{The long-run covariance estimator to be used for CUSUM method. Currently, only "Bartlett"and "Empirical" are supported. Alternatively, a matrix containing the long-run covariance estimate can be inputted.}

\item{thresholdValue}{Either the manual threshold value when threshold="Manual" or the (1-thresholdValue)-quantile of asymptotic distribution of the CUSUM test statistic when method="CUSUM" and threshold="Asymptotic".}

\item{errorCheck}{Logical. If TRUE error checking is performed.}

\item{Class}{Logical. If TRUE then an S4 class is returned. If FALSE the estimated changepoints are returned.}
}
\value{
An object of S4 class \code{\linkS4class{cptCovariance}} is returned. If Class="FALSE", the vector of changepoints are returned.
}
\description{
Implements the \insertCite{Aue2009;textual}{changepoint.cov} method for detecting covariance changes in multivariate time series. This method is aimed at low-dimensional time series that can have temporal dependence.
}
\details{
This function calculates the test statistic, T, described in \insertCite{Aue2009;textual}{changepoint.cov}, specifically the sum of the test statistics for all potential changepoint locations; changepoint localization uses the maximum of these test statistics. T is then normalized so that its asymptotic distribution is a standard Normal. This normalised test statistic is then compared to the defined threshold (either the specified quantile of the Normal distribution if the threshold is set as asymptotic or the manual threshold). If multiple changepoints are possible then the Binary Segmentation algorithm is used to detect multiple changes. The long run covariance estimation and its inversion can be unstable if the dimension of the time series is large. In this scenario we recommend using the \code{\link{cptRatio}} function.
}
\examples{
set.seed(1)
dataAMOC <- wishartDataGeneration(n=100,p=2,tau=50)$data

ansCUSUM <- cptCov(X=dataAMOC)
show(ansCUSUM)

ansCUSUM2 <- cptCov(X=dataAMOC,threshold='Manual',numCpts='AMOC',
		msl=15,thresholdValue=15,LRCov='Empirical')
summary(ansCUSUM2)

}
\references{
\insertRef{Aue2009}{changepoint.cov}
}
\seealso{
\code{\link{cptCov}}, \code{\linkS4class{cptCovariance}}, \code{\link{wishartDataGeneration}}
}
