% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cptFisher.R
\name{cptFisher}
\alias{cptFisher}
\title{Fisher matrix based covariance changepoint detection}
\usage{
cptFisher(
  X,
  threshold = "Asymptotic",
  noCpts = "AMOC",
  thresholdValue = NA,
  msl = dim(X)[2],
  Class = TRUE
)
}
\arguments{
\item{X}{Data matrix of dimension n by p}

\item{threshold}{Character threshold choice for determining significance of changepoints. Choices include:
\itemize{
\item "Asymptotic" - Uses the asymptotic distribution of the test statistic. See details for more information
\item "Manual"- A user chosen threshold which is contained in the thresholdValue parameter.
}}

\item{noCpts}{Number of changepoints in the data. Choices include:
\itemize{
\item "AMOC" - At Most One Changepoint; test to see if the data contains a single changepoint or not.
\item "BinSeg"- Binary segmentation is performed to detect multiple changepoints.
\item "Set" - Number of changepoints in the data is known and is contained in the m parameter. The method returns the m most significant changepoints via a Binary Segmentation framework.
}}

\item{thresholdValue}{User defined threshold when using threshold="Manual".}

\item{msl}{Minimum segment length allowed between the changepoints. NOTE this should be greater than or equal to p, the dimension of the time series.}

\item{Class}{Logical. If TRUE then an S4 class is returned. Else just the estimated changepoints are returned.}
}
\value{
An object of S4 class \code{\link{cptCov}} is returned. If Class="FALSE", just the vector of changepoints are returned.
}
\description{
Implements the Ryan method for detecting covariance changes in multivariate time series. This method is aimed at independent high-dimensional time series.
}
