% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cptRatio.R
\name{cptRatio}
\alias{cptRatio}
\title{Ratio method for covariance changepoint detection}
\usage{
cptRatio(
  X,
  threshold = "Asymptotic",
  numCpts = "AMOC",
  msl = 2 * ncol(X),
  thresholdValue = 0,
  errorCheck = TRUE,
  Class = TRUE
)
}
\arguments{
\item{X}{Data matrix of dimension n by p.}

\item{threshold}{Threshold choice for determining significance of changepoints. Choices include:
\itemize{
\item "Asymptotic" - Uses the asymptotic threshold derived for each method. For method="Ratio" the threshold is log(n). For method="CUSUM" the threshold is the specified quantile of the standard Normal distribution. The quantile is set by the argument thresholdValue.
\item "Manual"- A user chosen threshold which is contained in the thresholdValue argument. NOTE the normalized test statistics will be compared to the set thresholds - see details for more information.
}
If numCpts is numeric then the threshold is not used as the number of changepoints is known.}

\item{numCpts}{Number of changepoints in the data. Choices include:
\itemize{
\item "AMOC" - At Most One Changepoint; test to see if the data contains a single changepoint or not.
\item "BinSeg"- Binary segmentation is performed to detect multiple changepoints.
\item Numeric - User specified number of changepoints.
}}

\item{msl}{Minimum segment length allowed between the changepoints. NOTE this should be greater than or equal to p, the dimension of the time series.}

\item{thresholdValue}{Either the manual threshold value when threshold="Manual" or the (1-thresholdValue)-quantile of asymptotic distribution of the CUSUM test statistic when method="CUSUM" and threshold="Asymptotic".}

\item{errorCheck}{Logical. If TRUE error checking is performed}

\item{Class}{Logical. If TRUE then an S4 class is returned. If FALSE the estimated changepoints are returned.}
}
\value{
An object of S4 class \code{\link{cptCovariance-class}} is returned. If Class="FALSE", the vector of changepoints are returned.
}
\description{
Implements the \insertCite{Ryan2020;textual}{changepoint.cov} method for detecting covariance changes in multivariate time series. This method is aimed at independent high-dimensional time series.
}
\details{
This function calculates the test statistic, T, described in \insertCite{Ryan2020;textual}{changepoint.cov}. Using results from Random Matrix Theory the test statistic is normalised by it's asymptotic expectation and variance so that it follows a standard Normal distribution. Following the paper, the threshold \code{log(n)} is used if the threshold is set as asymptotic, else the user defined manual threshold is used. If multiple changepoints are possible then the Binary Segmentation algorithm is used to detect multiple changes. If the minimum segment length is too small then the numerical integration performed in the normalization of the test statistic can be unstable. In this scenario the minimum segment length will be automatically increased. This method is designed for independent time series, if the time series contains temporal dependence we recommend using the \code{\link{cptCUSUM}} function.
}
\examples{
set.seed(1)
dataAMOC <- wishartDataGeneration(n=100,p=5,tau=50)$data

ansRatio <- cptRatio(X=dataAMOC)
summary(ansRatio)

ansRatio2 <- cptCov(X=dataAMOC,threshold='Manual',numCpts='AMOC',msl=10,thresholdValue=20)
summary(ansRatio2)

}
\references{
\insertRef{Ryan2020}{changepoint.cov}
}
\seealso{
\code{\link{cptCov}}, \code{\linkS4class{cptCovariance}}, \code{\link{wishartDataGeneration}}, \code{\link{ratioTestStat}}
}
