% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cusumTestStat.R
\name{CusumCalculator}
\alias{CusumCalculator}
\title{CUSUM Calculator}
\usage{
CusumCalculator(X)
}
\arguments{
\item{X}{List of data where each slot is a time point}
}
\value{
A function used to calculate the CUSUM statistic
}
\description{
Creates a function that takes a changepoint location as an input and returns the CUSUM statistic assuming a changepoint at the given location.
}
