% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorChecks.R
\name{cusumErrorChecks}
\alias{cusumErrorChecks}
\title{Error checking for CUSUM Method}
\usage{
cusumErrorChecks(
  X,
  threshold,
  noCpts,
  LRCov,
  statType,
  thresholdValue,
  msl,
  Class
)
}
\description{
DEVELOPER USE ONLY. This function checks the user inputs to make sure they are all valid
}
