% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericErrorChecks.R
\name{dataErrorChecks}
\alias{dataErrorChecks}
\title{data error checking}
\usage{
dataErrorChecks(X)
}
\arguments{
\item{X}{Data matrix of dimension n by p.}
}
\description{
DEVELOPER USE ONLY. This function checks the user input for the data for all methods.
}
\seealso{
\code{\link{cptCov}}, \code{\link{cptSubspace}}
}
\keyword{internal}
