% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisherTestStat.R
\name{fisherDistanceCalculator}
\alias{fisherDistanceCalculator}
\title{Fisher distance calculator}
\usage{
fisherDistanceCalculator(X)
}
\arguments{
\item{X}{List of data where each slot is a time point}
}
\value{
A function used to calculate un-normalized test statistic
}
\description{
Creates a function that takes a potential changepoint location and returns the un-normalized test statistic defined in Ryan (2020).
}
