% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorChecks.R
\name{fisherErrorChecks}
\alias{fisherErrorChecks}
\title{Error checking for Fisher Method}
\usage{
fisherErrorChecks(X, threshold, noCpts, thresholdValue, msl, Class)
}
\description{
DEVELOPER USE ONLY. This function checks the user inputs to make sure they are all valid.
}
