% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratioTestStat.R
\name{fisherESD}
\alias{fisherESD}
\title{Empirical spectral distribution of a Fisher matrix}
\usage{
fisherESD(gamma1, gamma2)
}
\arguments{
\item{gamma1}{p/n1 where n1 is the number of time points before potential changepoint}

\item{gamma2}{p/n2 where n2 is the number of time points after potential changepoint}
}
\value{
A function that calculates the ESD of a Fisher matrix for given gammas
}
\description{
Creates a function which returns the ESD of a Fisher matrix with set gammas for some x within the support of the ESD
}
