% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratioTestStat.R
\name{fisherSupport}
\alias{fisherSupport}
\title{Calculates Fisher support}
\usage{
fisherSupport(gamma1, gamma2)
}
\arguments{
\item{gamma1}{p/n1 where n1 is the number of time points before potential changepoint}

\item{gamma2}{p/n2 where n2 is the number of time points after potential changepoint}
}
\value{
Support for Fisher ESD
}
\description{
Returns the support for calculating the Fisher ESD
}
