% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisherTestStat.R
\name{fisherTestStat}
\alias{fisherTestStat}
\title{Test statistic for Fisher method}
\usage{
fisherTestStat(X, msl)
}
\arguments{
\item{X}{Data matrix of dimension n by p}

\item{msl}{A numeric giving the minimum segment length between changepoints. NOTE this should be greater than or equal to p.}
}
\value{
A numeric vector containing the test statistic at each potential changepoint location.
}
\description{
Calculates the test statistic for all potential changepoint locations within the time series.
}
