% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratioTestStat.R
\name{functionProduct}
\alias{functionProduct}
\title{Product of two functions}
\usage{
functionProduct(f1, f2)
}
\arguments{
\item{f1}{Function 1}

\item{f2}{Function 2}
}
\value{
Function which is product of functions
}
\description{
Multiplies outputs of two functions
}
\seealso{
\code{\link{ratioTestStat}}
}
\keyword{internal}
