% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataGeneration.R
\name{generateTransMatrix}
\alias{generateTransMatrix}
\title{Transition matrix generation}
\usage{
generateTransMatrix(p, shape = 5, scale = 1/5)
}
\arguments{
\item{p}{Dimension of the time series.}

\item{shape}{Shape parameter for generation of eigenvalues in transition matrix. Only used with automatic generation of segment covariances.}

\item{scale}{Scale parameter for generation of eigenvalues in transition matrix. Only used with automatic generation of segment covariances.}
}
\value{
A p by p transition matrix
}
\description{
Generates transition matrices for \code{\link{wishartDataGeneration}}
}
\seealso{
\code{\link{wishartDataGeneration}}
}
\keyword{internal}
