% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericErrorChecks.R
\name{mslErrorChecks}
\alias{mslErrorChecks}
\title{msl error checking}
\usage{
mslErrorChecks(X, msl)
}
\arguments{
\item{X}{Data matrix of dimension n by p.}

\item{msl}{Minimum segment length allowed between the changepoints. NOTE this should be greater than or equal to p, the dimension of the time series.}
}
\description{
DEVELOPER USE ONLY. This function checks the user input for the msl argument for all methods.
}
