% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericErrorChecks.R
\name{numCptsErrorChecks}
\alias{numCptsErrorChecks}
\title{numCpts error checking}
\usage{
numCptsErrorChecks(numCpts, method)
}
\arguments{
\item{numCpts}{Number of changepoints in the data. Choices include:
\itemize{
\item "AMOC" - At Most One Changepoint; test to see if the data contains a single changepoint or not.
\item "BinSeg"- Binary segmentation is performed to detect multiple changepoints.
\item Numeric - User specified number of changepoints.
}}

\item{method}{Covariance changepoint method to be used. Choice of "Ratio" or "CUSUM".}
}
\description{
DEVELOPER USE ONLY. This function checks the user input for the numCpts argument for all methods.
}
\seealso{
\code{\link{cptCov}}, \code{\link{cptSubspace}}
}
\keyword{internal}
