% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratioTestStat.R
\name{ratioTestStat}
\alias{ratioTestStat}
\title{Test statistic for Ratio method}
\usage{
ratioTestStat(X, msl)
}
\arguments{
\item{X}{Data matrix of dimension n by p.}

\item{msl}{A numeric giving the minimum segment length between changepoints. NOTE this should be greater than or equal to p.}
}
\value{
A numeric vector containing the test statistic at each potential changepoint location.
}
\description{
Calculates the Ratio test statistic for all potential changepoint locations within the time series. See \code{\link{cptRatio}} for more details.
}
\details{
See \code{\link{cptRatio}}.
}
\examples{
set.seed(1)
data <- wishartDataGeneration(n=100,p=3,tau=50)$data
ans <- ratioTestStat(X=data,msl=20)
which.max(ans)

}
\references{
\insertRef{Ryan2020}{changepoint.cov}
}
\seealso{
\code{\link{cptRatio}}, \code{\link{cptCov}}
}
