% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cptCovarianceClass.R
\name{replacementGeneric}
\alias{replacementGeneric}
\alias{dataset<-}
\alias{cptsSig<-}
\alias{cpts<-}
\alias{method<-}
\alias{msl<-}
\alias{numCpts<-}
\alias{threshold<-}
\alias{thresholdValue<-}
\alias{subspaceDim<-}
\alias{nperm<-}
\alias{LRCov<-}
\title{Replacement Functions - Generic}
\usage{
dataset(x) <- value

cptsSig(x) <- value

cpts(x) <- value

method(x) <- value

msl(x) <- value

numCpts(x) <- value

threshold(x) <- value

thresholdValue(x) <- value

subspaceDim(x) <- value

nperm(x) <- value

LRCov(x) <- value
}
\arguments{
\item{x}{object of class \code{\linkS4class{cptCovariance}}}

\item{value}{value}
}
\description{
Replacement Functions - Generic
}
