% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cptCovariance-class.R
\name{replacementMethod}
\alias{replacementMethod}
\alias{dataset<-,cptCovariance-method}
\alias{cptsSig<-,cptCovariance-method}
\alias{cpts<-,cptCovariance-method}
\alias{method<-,cptCovariance-method}
\alias{msl<-,cptCovariance-method}
\alias{numCpts<-,cptCovariance-method}
\alias{threshold<-,cptCovariance-method}
\alias{thresholdValue<-,cptCovariance-method}
\alias{subspaceDim<-,cptCovariance-method}
\alias{nperm<-,cptCovariance-method}
\alias{LRCov<-,cptCovariance-method}
\title{Replacement Functions - Method}
\usage{
\S4method{dataset}{cptCovariance}(x) <- value

\S4method{cptsSig}{cptCovariance}(x) <- value

\S4method{cpts}{cptCovariance}(x) <- value

\S4method{method}{cptCovariance}(x) <- value

\S4method{msl}{cptCovariance}(x) <- value

\S4method{numCpts}{cptCovariance}(x) <- value

\S4method{threshold}{cptCovariance}(x) <- value

\S4method{thresholdValue}{cptCovariance}(x) <- value

\S4method{subspaceDim}{cptCovariance}(x) <- value

\S4method{nperm}{cptCovariance}(x) <- value

\S4method{LRCov}{cptCovariance}(x) <- value
}
\arguments{
\item{x}{object of class \code{\linkS4class{cptCovariance}}}

\item{value}{value}
}
\description{
Replacement Functions - Method
}
