% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cptCovarianceClass.R
\name{retrievalGeneric}
\alias{retrievalGeneric}
\alias{dataset}
\alias{cptsSig}
\alias{cpts}
\alias{method}
\alias{msl}
\alias{numCpts}
\alias{threshold}
\alias{thresholdValue}
\alias{subspaceDim}
\alias{nperm}
\alias{LRCov}
\title{Retrieval Functions - Generic}
\usage{
dataset(x)

cptsSig(x)

cpts(x)

method(x)

msl(x)

numCpts(x)

threshold(x)

thresholdValue(x)

subspaceDim(x)

nperm(x)

LRCov(x)
}
\arguments{
\item{x}{object of class \code{\linkS4class{cptCovariance}}}
}
\description{
Retrieval Functions - Generic
}
