% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cptCovarianceClass.R
\name{retrievalMethod}
\alias{retrievalMethod}
\alias{dataset,cptCovariance-method}
\alias{cptsSig,cptCovariance-method}
\alias{cpts,cptCovariance-method}
\alias{method,cptCovariance-method}
\alias{msl,cptCovariance-method}
\alias{numCpts,cptCovariance-method}
\alias{threshold,cptCovariance-method}
\alias{thresholdValue,cptCovariance-method}
\alias{subspaceDim,cptCovariance-method}
\alias{nperm,cptCovariance-method}
\alias{LRCov,cptCovariance-method}
\title{Retrieval Functions - Method}
\usage{
\S4method{dataset}{cptCovariance}(x)

\S4method{cptsSig}{cptCovariance}(x)

\S4method{cpts}{cptCovariance}(x)

\S4method{method}{cptCovariance}(x)

\S4method{msl}{cptCovariance}(x)

\S4method{numCpts}{cptCovariance}(x)

\S4method{threshold}{cptCovariance}(x)

\S4method{thresholdValue}{cptCovariance}(x)

\S4method{subspaceDim}{cptCovariance}(x)

\S4method{nperm}{cptCovariance}(x)

\S4method{LRCov}{cptCovariance}(x)
}
\arguments{
\item{x}{object of class \code{\linkS4class{cptCovariance}}}
}
\description{
Retrieval Functions - Method
}
