% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subspaceTestStatistic.R
\name{subspaceCost}
\alias{subspaceCost}
\title{Cost of data for Subspace method}
\usage{
subspaceCost(X, q)
}
\arguments{
\item{X}{Data matrix of dimension n by p}

\item{q}{Dimension of the latent subspace}
}
\value{
Cost of data
}
\description{
Calculates the cost of data assuming a latent subspace of dimension q
}
