% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subspaceTestStatistic.R
\name{subspaceCostCalculator}
\alias{subspaceCostCalculator}
\title{Subspace cost calculator}
\usage{
subspaceCostCalculator(X, q)
}
\arguments{
\item{X}{Data matrix of dimension n by p}

\item{q}{Dimension of the latent subspace}
}
\value{
A function used to calculate the test statistic
}
\description{
Creates a function that takes potential changepoint location and returns the test statistic as defined in Grundy(2020)
}
