% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subspaceErrorChecks.R
\name{subspaceErrorChecks}
\alias{subspaceErrorChecks}
\title{Error checking for Subspace Method}
\usage{
subspaceErrorChecks(X, q, threhsold, noCpts, thresholdValue, msl, nperm, m)
}
\description{
DEVELOPER USE ONLY. This function checks the user inputs to make sure they are all valid
}
