% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subspaceTestStatistic.R
\name{subspaceTestStat}
\alias{subspaceTestStat}
\title{Test statistic for Subspace method}
\usage{
subspaceTestStat(X, q, msl = dim(X)[2])
}
\arguments{
\item{X}{Data matrix of dimension n by p}

\item{q}{Dimension of the latent subspace}

\item{msl}{Minimum segment length between changepoints. NOTE this should be greater than or equal to p}
}
\value{
A numeric vector containing the test statistic at each potential changepoint location
}
\description{
Calculates the Subspace test statistic for all potential changepoint locations within the time series.
}
