% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericErrorChecks.R
\name{thresholdErrorChecks}
\alias{thresholdErrorChecks}
\title{threshold error checking}
\usage{
thresholdErrorChecks(threshold, thresholdValue, method)
}
\arguments{
\item{threshold}{Type of threshold to be used}

\item{thresholdValue}{Numeric value of threshold or quantile of asymptotic distribution/permutation test}

\item{method}{function name for relevant help file}
}
\description{
DEVELOPER USE ONLY. This function checks the user input for the threshold and thresholdValue.
}
\seealso{
\code{\link{cptCov}}, \code{\link{cptSubspace}}
}
\keyword{internal}
