% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cusumTestStat.R
\name{vech}
\alias{vech}
\title{Column stacking}
\usage{
vech(X)
}
\arguments{
\item{X}{A data matrix of dimension n by p}
}
\value{
A numeric vector of length p(p+1)/2
}
\description{
Stacks the columns below the diagonal of a symmetric matrix
}
