% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binSeg.R
\name{binSeg}
\alias{binSeg}
\title{Binary Segmentation Implementation}
\usage{
binSeg(
  X,
  method,
  msl,
  threshold,
  thresholdValue,
  m = -1,
  LRCov = "Bartlett",
  q = 1,
  nperm = 200,
  alpha = 0.05
)
}
\arguments{
\item{X}{Data matrix of dimension n by p.}

\item{method}{Choice of "Ratio", "CUSUM" or "Subspace".}

\item{msl}{Minimum segment length}

\item{threshold}{Threshold choice. Only relevant for using the permutation test within the Subspace method. If method="Subspace" and threshold="PermTest" then the permutation test is used to generate a threshold at each round of the binary segmentation procedure.}

\item{thresholdValue}{The threshold value that needs determines if a changepoint is significant}

\item{m}{User specified number of changepoints. Default is m=-1 meaning the thresholdValue is used as a stopping criteria}

\item{LRCov}{The long-run covariance estimator to be used for CUSUM method. Currently, only "Bartlett" and "Empirical" are supported.}

\item{q}{Dimension of the latent subspace for Subspace method.}

\item{nperm}{Only required for Subspace method when threshold="PermTest". Number of permutations to use in the permutation test.}

\item{alpha}{Only required for Subspace method when threshold="PermTest". Significance level of the permutation test.}
}
\value{
A data frame containing the identified changepoints along with their test statistic and threshold they exceeded.
}
\description{
This function estimates multiple covariance changepoints. NOTE no error checking is performed. This function is for developer use only.
}
