% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classInput.R
\name{classInput}
\alias{classInput}
\title{Inputs results into S4 class}
\usage{
classInput(
  X,
  cpts,
  method,
  noCpts,
  testStat,
  threshold,
  thresholdValue,
  msl,
  q = 0,
  nperm = 0,
  LRCov = "NA",
  statType = "NA"
)
}
\description{
Function for inputting results into the S4 class \code{\link{cptCov}}.
}
