% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cptCovarianceClass.R
\docType{class}
\name{cptCovariance-class}
\alias{cptCovariance-class}
\alias{summary,cptCovariance-method}
\alias{show,cptCovariance-method}
\alias{data,cptCovariance-method}
\alias{cpts,cptCovariance-method}
\alias{method,cptCovariance-method}
\alias{msl,cptCovariance-method}
\alias{numCpts,cptCovariance-method}
\alias{threshold,cptCovariance-method}
\alias{thresholdValue,cptCovariance-method}
\alias{cptsSig,cptCovariance-method}
\alias{subspaceDim,cptCovariance-method}
\alias{nperm,cptCovariance-method}
\alias{LRCov,cptCovariance-method}
\title{An S4 class for a covariance changepoint object}
\usage{
\S4method{summary}{cptCovariance}(object)

\S4method{show}{cptCovariance}(object)

\S4method{data}{cptCovariance}(object)

\S4method{cpts}{cptCovariance}(object)

\S4method{method}{cptCovariance}(object)

\S4method{msl}{cptCovariance}(object)

\S4method{numCpts}{cptCovariance}(object)

\S4method{threshold}{cptCovariance}(object)

\S4method{thresholdValue}{cptCovariance}(object)

\S4method{cptsSig}{cptCovariance}(object)

\S4method{subspaceDim}{cptCovariance}(object)

\S4method{nperm}{cptCovariance}(object)

\S4method{LRCov}{cptCovariance}(object)
}
\arguments{
\item{object}{An object of S4 class \code{\link{cptCovariance-class}}}
}
\description{
An S4 class for a covariance changepoint object
}
\section{Methods (by generic)}{
\itemize{
\item \code{summary}: Summarises the cptCovariance object

\item \code{show}: Shows the cptCovariance object

\item \code{data}: Retrieves the data slot

\item \code{cpts}: Retrieves the cpts slot

\item \code{method}: Retrieves the method slot

\item \code{msl}: Retrieves the msl slot

\item \code{numCpts}: Retrieves the numCpts slot

\item \code{threshold}: Retrieves the threshold slot

\item \code{thresholdValue}: Retrieves the thresholdValue slot

\item \code{cptsSig}: Retrieves the cptsSig slot

\item \code{subspaceDim}: Retrieves the subspaceDim slot

\item \code{nperm}: Retrieves the nperm slot

\item \code{LRCov}: Retrieves the LRCov slot
}}

\section{Slots}{

\describe{
\item{\code{data}}{An n by p matrix of the data}

\item{\code{cpts}}{A numeric vector containing the identified changepoints}

\item{\code{method}}{Character containing covariance changepoint method used}

\item{\code{numCpts}}{Either 'AMOC' for at most one changepoint; 'BinSeg' for a binary segmentation approach to dect multiple changepoints; or a positive integer specifying the number of changes}

\item{\code{cptsSig}}{Data frame containing the changepoint locations along with their associated test statistic; threshold; and whether or not they were deemed significant}

\item{\code{threshold}}{Character containing the method used for generating the threshold}

\item{\code{thresholdValue}}{Threshold value used to determine significant changepoints. If permutation test within subspace method is used then a vector of threshold values is returned.}

\item{\code{msl}}{Numeric containing the minimum segment length between changepoints}

\item{\code{subspaceDim}}{Numeric value of subspace dimension. Only used for Subspace method}

\item{\code{nperm}}{Numeric value of number of permutations used in permutation test. Only used for subspace method when threshold is "PermTest"}

\item{\code{LRCov}}{Character describing the long-run covariance estimator used. Only used for Aue method.}

\item{\code{date}}{Creation date of the object}

\item{\code{version}}{Version of the cpt.covariance used}
}}

\examples{
ans <- new('cptCovariance',data=matrix(rnorm(300),ncol=3),
		cpts=c(50,100),
		method='Ratio',
		numCpts='AMOC',
		cptsSig=data.frame('cpts'=50,'T'=33.3,'thresholdValue'=30,significant=TRUE),
		threshold='Manual',
		thresholdValue=30,
		msl=20)
summary(ans)
show(ans)

}
