% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cptCUSUM.R
\name{cptCUSUM}
\alias{cptCUSUM}
\title{CUSUM based covariance changepoint detection}
\usage{
cptCUSUM(
  X,
  threshold = "Normal",
  noCpts = "AMOC",
  LRCov = "Bartlett",
  statType = "Mean",
  thresholdValue = 0.05,
  msl = dim(X)[2],
  Class = TRUE
)
}
\arguments{
\item{X}{Data matrix of dimension n by p}

\item{threshold}{Character threshold choice for determining significance of changepoints. Choices include:
\itemize{
\item "Normal" - Uses the normality of the asymptotic distribution of the test statistic. The significance level is contained in the thresholdValue parameter.
\item "Manual"- A user chosen threshold which is contained in the thresholdValue parameter. NOTE the normalized test statistic will be compared to the set threshold.
}}

\item{noCpts}{Number of changepoints in the data. Choices include:
\itemize{
\item "AMOC" - At Most One Changepoint; test to see if the data contains a single changepoint or not.
\item "BinSeg"- Binary segmentation is performed to detect multiple changepoints.
\item "Set" - Number of changepoints in the data is known and is contained in the m parameter. The method returns the m most significant changepoints via a Binary Segmentation framework.
}}

\item{LRCov}{Character stating the long-run covariance estimator to be used. Currently, only "Bartlett" and "Empirical" are supported.}

\item{statType}{Character stating which test statistic should be used. Choice of "Mean" and "Max".}

\item{thresholdValue}{Either the significance level of the Normal distribution when using threshold="Normal" or the user defined threshold when using threshold="Manual".}

\item{msl}{Minimum segment length allowed between the changepoints. NOTE this should be greater than or equal to p, the dimension of the time series.}

\item{Class}{Logical. If TRUE then an S4 class is returned. Else just the estimated changepoints are returned.}
}
\value{
An object of S4 class \code{\link{cptCov}} is returned. If Class="FALSE", just the vector of changepoints are returned.
}
\description{
Implements the Aue method for detecting covariance changes in multivariate time series. This method is aimed at low-dimensional time series that can have temporal dependence.
}
