% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cptSubspace.R
\name{cptSubspace}
\alias{cptSubspace}
\title{Detecting changes in subspace}
\usage{
cptSubspace(
  X,
  q,
  threshold = "PermTest",
  noCpts = "AMOC",
  thresholdValue = 0.05,
  msl = dim(X)[2],
  nperm = 200,
  m = NA,
  Class = FALSE
)
}
\arguments{
\item{X}{Data matrix of dimension n by p}

\item{q}{Dimension of the latent subspace}

\item{threshold}{Threshold choice for determining significance of changepoints. Choices include:
\itemize{
\item "PermTest" - Permutation test is performed using the number of permutations and significance level contained in the nperm and thresholdValue paramters respectively
\item "Manual" - A user chosen threshold is used which is contained in the thresholdValue parameter.
}}

\item{noCpts}{Number of changepoints in the data. Choices inlcude:
\itemize{
\item "AMOC" - At Most One Changepoint; test to see if the data contains a single changepoint or not.
\item "BinSeg"- Binary segmentation is performed to detect multiple changepoints.
\item "Set" - Number of changepoints in the data is known and is contained in the m parameter. The method returns the m most significant changepoints via a Binary Segmentation framework.
}}

\item{thresholdValue}{Either the significance level of the permutation test when using threshold="PermTest" or the user defined threshold when using threshold="Manual".}

\item{msl}{Minimum segment length allowed between the changepoints. NOTE this should be greater than or equal to p, the dimension of the time series.}

\item{nperm}{Only required for threshold="PermTest". Number of permutations to use in the permutation test.}

\item{m}{Only required for noCpts="Set". Known number of changepoints in the data.}

\item{Class}{Logical. If TRUE then an S4 class is returned. Else just the estimated changepoints are returned.}
}
\value{
An object of S4 class "cptCovSubspace" is returned. If Class="FALSE", just the vector of changepoints are returned.
}
\description{
This method allows the detection of subspace changes in multivariate time series data.
}
