% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cusumTestStat.R
\name{cusumTestStat}
\alias{cusumTestStat}
\title{Test statistic for CUSUM method}
\usage{
cusumTestStat(X, LRCov, msl)
}
\arguments{
\item{X}{Data matrix of dimension n by p}

\item{LRCov}{A character describing the Long-run covariance estimator to be used. Currently, only "Bartlett" and "Empirical" are available, see \code{\link{cptAue}} for more details.}

\item{msl}{A numeric giving the minimum segment length between changepoints. NOTE this should be greater than or equal to p.}
}
\value{
A numeric vector containing the test statistic at each potential changepoint location.
}
\description{
Calculates the test statistic for all potential changepoint locations within the time series.
}
