% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutationTest.R
\name{permutationTest}
\alias{permutationTest}
\title{Permutation Test for Subspace Method}
\usage{
permutationTest(X, q, msl = dim(X)[2], alpha = 0.05, nperm = 200)
}
\arguments{
\item{X}{Data matrix of dimension n by p}

\item{q}{Dimension of the latent subspace}

\item{msl}{Minimum segment length between changepoints. Note this should be greater than or equal to p}

\item{alpha}{Significance level of test}

\item{nperm}{Number of random permutations to be performed}
}
\value{
Numeric containing the threshold value
}
\description{
Calculates a threshold for the Subspace Method via a permutation test.
}
