% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodErrorChecks.R
\name{ratioErrorChecks}
\alias{ratioErrorChecks}
\title{Error checking for Ratio Method}
\usage{
ratioErrorChecks(X, threshold, numCpts, thresholdValue, msl, Class)
}
\arguments{
\item{X}{Data matrix of dimension n by p.}

\item{threshold}{Threshold choice for determining significance of changepoints. Choices include:
\itemize{
\item "Asymptotic" - Uses the asymptotic threshold derived for each method. For method="Ratio" the threshold is log(n). For method="CUSUM" the threshold is the specified quantile of the standard Normal distribution. The quantile is set by the argument thresholdValue.
\item "Manual"- A user chosen threshold which is contained in the thresholdValue argument. NOTE the normalized test statistics will be compared to the set thresholds - see details for more information.
}
If numCpts is numeric then the threshold is not used as the number of changepoints is known.}

\item{numCpts}{Number of changepoints in the data. Choices include:
\itemize{
\item "AMOC" - At Most One Changepoint; test to see if the data contains a single changepoint or not.
\item "BinSeg"- Binary segmentation is performed to detect multiple changepoints.
\item Numeric - User specified number of changepoints.
}}

\item{thresholdValue}{Either the manual threshold value when threshold="Manual" or the (1-thresholdValue)-quantile of asymptotic distribution of the CUSUM test statistic when method="CUSUM" and threshold="Asymptotic".}

\item{msl}{Minimum segment length allowed between the changepoints. NOTE this should be greater than or equal to p, the dimension of the time series.}

\item{Class}{Logical. If TRUE then an S4 class is returned. If FALSE the estimated changepoints are returned.}
}
\description{
DEVELOPER USE ONLY. This function checks the user inputs to make sure they are all valid.
}
\seealso{
\code{\link{cptRatio}}
}
\keyword{internal}
