% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subspaceTestStat.R
\name{subspaceCostCalculator}
\alias{subspaceCostCalculator}
\title{Subspace cost calculator}
\usage{
subspaceCostCalculator(X, subspaceDim)
}
\arguments{
\item{X}{Data matrix of dimension n by p}

\item{subspaceDim}{Dimension of the latent subspace}
}
\value{
A function used to calculate the test statistic
}
\description{
Creates a function that takes potential changepoint location and returns the
test statistic as defined in \insertCite{Grundy2020;textual}{changepoint.cov}.
}
\seealso{
\code{\link{subspaceTestStat}}
}
\keyword{internal}
