% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataGeneration.R
\name{subspaceDataGeneration}
\alias{subspaceDataGeneration}
\title{Data Generation - Subspace}
\usage{
subspaceDataGeneration(
  n,
  p,
  q,
  tau = c(1, n),
  nvar = 0.05,
  svar = 1,
  changeSize = 0.5 * sqrt(q)
)
}
\arguments{
\item{n}{Number of time points}

\item{p}{Dimension of the time series}

\item{q}{Dimension of the latent subspace}

\item{tau}{Vector of changepoint locations}

\item{nvar}{Variance of measurement error (assumed to be Normal)}

\item{svar}{Variance within the subspace i.e. the variance of points in directions of the basis of the subspace (assumed to be Normal)}

\item{changeSize}{Either scalar for change size of all the changepoints or a vector of equal length to the number of changepoints indicating the change size for each changepoint. All change sizes must be between 0 and sqrt(q).}
}
\value{
List with elements:
\itemize{
\item data Matrix of generated data with dimension n by p
\item cpts Location of the changepoints
\item changeSize Scalar or vector of change sizes
\item subspace List of the bases of the subspace for each segment
}
}
\description{
Allows for easy generation of data that lies in a latent subspace with changepoints in the subspace.
}
\examples{
set.seed(1)
dataObject <- subspaceDataGeneration(n=100,p=4,q=2,tau=40)
dim(dataObject$data)
dataObject$cpts
dataObject$changeSize
dataObject$subspace

}
