% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subspaceTestStatistic.R
\name{subspaceTestStat}
\alias{subspaceTestStat}
\title{Test statistic for Subspace method}
\usage{
subspaceTestStat(X, q, msl = dim(X)[2])
}
\arguments{
\item{X}{Data matrix of dimension n by p}

\item{q}{Dimension of the latent subspace}

\item{msl}{Minimum segment length between changepoints. NOTE this should be greater than or equal to p}
}
\value{
A list with slots:
\itemize{
\item T - Test statistic
\item cpt - Most likely changepoint position
\item nullCost - Cost of data with no changes
\item altCost - List with cost of data with a changepoint at the slot location
}
}
\description{
Calculates the test statistic, most likely changepoint location and the null and alternative costs for the Subspace Covariance changepoint method.
}
