% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subspaceTestStatistic.R
\name{subspaceTestStat}
\alias{subspaceTestStat}
\title{Test statistic for Subspace method}
\usage{
subspaceTestStat(X, q, msl)
}
\arguments{
\item{X}{Data matrix of dimension n by p}

\item{q}{Dimension of the latent subspace}

\item{msl}{Minimum segment length between changepoints. NOTE this should be greater than or equal to p}
}
\value{
A numeric vector containing the test statistic at each potential changepoint location
}
\description{
Calculates the Subspace test statistic for all potential changepoint locations within the time series.
}
\examples{
set.seed(1)
data <- subspaceDataGeneration(n=100,p=20,q=5,tau=50,changeSize=0.5*sqrt(5))$data
ans <- subspaceTestStat(X=data,q=5,msl=30)
which.max(ans)

}
